src/server/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
websockets>=12.0
httpx>=0.27.0
python-multipart>=0.0.9
aiofiles>=23.2.1
python-dotenv>=1.0.0
